requirements.txt
asgiref==3.2.3
Django==3.0.1
django-mathfilters==0.4.0
django-tinymce4-lite==1.7.5
jsmin==2.2.2
mysqlclient==1.4.6
pytz==2019.3
sqlparse==0.3.0
